Move definitions of Nulecule and NuleculeComponent to nulecule/base.py
